refactored as separate component, reset TextInput
